taskly/requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
cffi==1.17.1
cryptography==45.0.7
Django==5.2.6
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-ninja==1.4.3
pillow==11.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-decouple==3.8
setuptools==80.9.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
